cleaned up code, modified profile/editprofile pages
