+++ src/pages/ProfilePage/ProfilePage.css
@@ -21,6 +21,7 @@ body {
 
 .profile-picture {
   margin-bottom: 10px;
+  margin-top: 75px;
 }
 
 .profile-name {
